LayoutTests/media/video-seek-past-end-playing-expected.txt

Test that seeking video with 'loop' past it's end rewinds to the beginning and continues playback.

RUN(video.src = 'content/test.mp4')

EVENT(canplaythrough)
EXPECTED (video.paused == 'true') OK
RUN(video.play())
EXPECTED (video.paused == 'false') OK

EXPECTED (video.paused == 'false') OK
EXPECTED (mediaElement.currentTime > '0') OK
RUN(video.currentTime = 500)

EXPECTED (video.paused == 'false') OK
EXPECTED (mediaElement.currentTime > '0') OK

END OF TEST
